% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neighbours.R
\name{get_neighbours}
\alias{get_neighbours}
\title{get_neighbours}
\usage{
get_neighbours(df, plot_ID)
}
\arguments{
\item{plot_ID}{}
}
\value{

}
\description{
Function to compute the RDPI (Relative Distance Plasticity Index,
Valladares et al, (2006) Quantitative estimation of phenotypic plasticity:
 bridging the gap between the evolutionary concept and its ecological applications,
 Journal of Ecology, 94(6):1103-1116.
}
\examples{
plot <-c(rep(1:2, 9), rep(3:6, 14), rep(7,6))
sps_pool <- c("PINI", "PISY", "ABAL")
sps <- sample(sps_pool, length(plot), replace = T)
dbh <- rnorm(length(plot), 15, 5)
x <- rnorm(length(plot), 0, 5)
y <- rnorm(length(plot), 0, 5)
data <- data.frame(plot, sps, dbh, x, y)
neighbours <- get_neighbours(data, plot)


}
