% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\docType{data}
\name{neighbors}
\alias{neighbors}
\title{Forest inventory data in Pyrenean forests}
\format{
A data frame with columns:
\describe{
 \item{plot}{Plot identifier}
 \item{sps_target}{Target tree species, with 3 possible values: ABAL, PINI or PISY}
 \item{dbh_target}{Target tree diameter(cm).}
 \item{x_target}{x coordinates of the target tree within the plot, as measured from the plot center (m).}
 \item{y_target}{y coordinates of the target tree within the plot, as measured from the plot center (m).}
 \item{n_target}{Correlative number identifying target tree position within the plot.}
 \item{ID_target}{Unique identifier for target trees}
 \item{sps_neighbor}{Neighbor tree species, with 3 possible values: ABAL, PINI or PISY}
 \item{dbh_neighbor}{Neighbor tree diameter(cm).}
 \item{x_neighbor}{x coordinates of the neighbor tree within the plot, as measured from the plot center (m).}
 \item{y_neighbor}{y coordinates of the neighbor tree within the plot, as measured from the plot center (m).}
 \item{n_neighbor}{Correlative number identifying neighbor position within the plot.}
 \item{ID_neighbor}{Unique identifier for neighbor trees}
}
}
\source{
Aitor Ameztegui
}
\usage{
neighbors
}
\description{
Forest inventory data in Pyrenean forests
}
\examples{
\dontrun{
 neighbors
}
}
\keyword{datasets}
