% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_data.R
\docType{data}
\name{tree_data}
\alias{tree_data}
\title{Forest inventory data in Pyrenean forests}
\format{
A data frame with columns:
\describe{
 \item{plot}{Plot identifier}
 \item{sps}{Tree species, with 3 possible values: ABAL, PINI or PISY}
 \item{dbh}{Tree diameter(cm).}
 \item{x}{x coordinates of the tree within the plot, as measured from the plot center (m).}
 \item{y}{y coordinates of the tree within the plot, as measured from the plot center (m).}
}
}
\source{
Aitor Ameztegui
}
\usage{
tree_data
}
\description{
Forest inventory data in Pyrenean forests
}
\examples{
\dontrun{
 tree_data
}
}
\keyword{datasets}
