% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_nt_tracts2022}
\alias{us_nt_tracts2022}
\title{US census tract neighborhood racial typologies, 2022}
\format{
See [us_nt_tracts2019]; columns are identical.
}
\source{
US Census Bureau, American Community Survey 5-year (table B03002).
}
\usage{
us_nt_tracts2022
}
\description{
Tract-level racial/ethnic composition and `ntdf()` typology assignments for
all US states using 5-year ACS data centered on 2022.
}
\keyword{datasets}
