% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afford.R
\name{afford}
\alias{afford}
\title{Affordable market location}
\usage{
afford(
  state = "state",
  counties = "counties",
  ami_limit = "ami_limit",
  year = 2019,
  ...
)
}
\arguments{
\item{state}{Study state.}

\item{counties}{Study county or counties.}

\item{ami_limit}{Specified area median income, or AMI, limit of interest such as 3. or .5 AMI.}

\item{year}{Study year.}

\item{...}{Other keyword arguments}
}
\value{
Returns a spatial file
}
\description{
The \code{afford} function identifies where different income groups can afford to live based on local market value while accounting for the region's count of households of the given income group.
}
\examples{
\dontrun{
bay5 <- afford("06", c("001", "013", "041", "075", "081"), .5, 2019)
}
}
