% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nci_files.R
\name{create_nci_files}
\alias{create_nci_files}
\title{create_nci_files}
\usage{
create_nci_files(df, plot_ID, var)
}
\arguments{
\item{df}{the dataframe containing information about the target trees and their neighbors. It can be provided
by the users or obtained using the `get_neighbors` function}

\item{plot_ID}{optional. Variable that identifies the plots, or experimental units, within which
the neighbors will be considered. This information is used to split the calculations per plot, and the results
are merged back into a single data frame.}

\item{var}{The variable that we want to extract from the neighbors. It can be either a numeric, logical or
character variable.}
}
\value{
a data frame containing as many rows as target trees. Each row will contain as many values (columns)
as neighbors has the corresponding target tree. Cells contain values of the variable "var" for each neighbor tree, and
missing values are coded as NA
}
\description{
function to create the files needed to estimate NCI-growth euqations using
the anneal function of the likelihood package.
}
\details{
create_nci_files
}
\examples{

data(neighbors)

dbhs <- create_nci_files(neighbors, plot, dbh_neighbor)
sps <- create_nci_files(neighbors, plot, sps_neighbour)
distances <- create_nci_files(neighbors, plot, dist)


}
