% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nci_files.R
\name{create_nci_files}
\alias{create_nci_files}
\title{get_neighbours}
\usage{
create_nci_files(df, plot_ID, var)
}
\arguments{
\item{var}{}
}
\value{

}
\description{
Function to compute the RDPI (Relative Distance Plasticity Index,
Valladares et al, (2006) Quantitative estimation of phenotypic plasticity:
 bridging the gap between the evolutionary concept and its ecological applications,
 Journal of Ecology, 94(6):1103-1116.
}
\details{
Title
}
\examples{


}
