% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neighbor_variables.R
\name{get_neighbor_variables}
\alias{get_neighbor_variables}
\title{get_neighbor_variables}
\usage{
get_neighbor_variables(df, plot_ID, var, self_del = T)
}
\arguments{
\item{df}{the data frame containing information about the target trees.}

\item{plot_ID}{optional. Variable that identifies the plots, or experimental units, within which
the neighbors will be considered. This information is used to split the calculations per plot, and the results
are merged back into a single data frame. A tree `b` will not be considered as a neighbor of tree `a` unless they belong to the same
`plot_ID`. If this argument is missing, all trees in `df` will be considered both targets and neighbors.}

\item{var}{The variable that we want to extract from the neighbors. It can be either a numeric, logical or
character variable.}

\item{self_del}{If `TRUE`, a tree is not included as a neighbor of itself. If `FALSE` the attribute if the target tree will be also
included in the output as a neighbor of itself.}
}
\value{
a data frame containing as many rows as target trees, and as many columns as the maximum number of neighbors per tree.
Cells contain values of the variable "var" for each neighbor tree, and missing values are coded as `NA`.
}
\description{
function to obtain the attributes of all the neighbors of a given set of target trees
}
\details{
get_neighbor_variables
}
\examples{

data(tree_data)

# Get dbhs of all the neighbirs
dbhs <-get_neighbor_variables(tree_data, plot, dbh, self_del = T)

# Get species of all the neighbors (incuding target tree as its own neighbor)
species <-get_neighbor_variables(tree_data, plot, sps, self_del = F)

}
