% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nt_pal.R
\name{nt_pal}
\alias{nt_pal}
\title{Suggested Color Palette}
\usage{
nt_pal(df, ...)
}
\arguments{
\item{df}{ntdf object}

\item{...}{Other keyword arguments}
}
\value{
Returns a factored color palette
}
\description{
Suggested Color Palette
}
