% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntcheck.R
\name{ntcheck}
\alias{ntcheck}
\title{Investigate possible concatenation}
\usage{
ntcheck(df, ...)
}
\arguments{
\item{df}{\code{ntdf} created dataframe}

\item{...}{Other keyword arguments}
}
\value{
Frequncy table of each neighborhood type
}
\description{
Investigate possible concatenation
}
\examples{
\dontrun
}
