% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntdf.R
\name{ntdf}
\alias{ntdf}
\title{Create segregation measure}
\usage{
ntdf(
  state,
  geography = "tract",
  county = NULL,
  geometry = FALSE,
  cache_table = TRUE,
  output = "wide",
  year = 2019,
  ...
)
}
\arguments{
\item{state}{Study state or states}

\item{geography}{Default "tract", see tidycensus for options}

\item{county}{Study county or counties}

\item{geometry}{Download sf spatial feature. Default is FALSE}

\item{cache_table}{Default TRUE}

\item{output}{Default "wide", other option is "tidy"}

\item{year}{Default "NULL"}

\item{...}{Other keyword arguments}
}
\value{
A dataframe of racial segregated neighborhoods
}
\description{
Drawn from Hall, Matthew, Kyle Crowder, and Amy Spring. 2015. “Neighborhood Foreclosures, Racial/Ethnic Transitions, and Residential Segregation.” American Sociological Review 80:526–549.
}
\examples{
\dontrun
Baltimore_nt <- ntdf(state = "MD", county = "Baltimore City", geometry = TRUE)
dplyr::glimpse(Baltimore_nt)
}
